Split a test
