+++ src/axom/bump/tests/CMakeLists.txt
@@ -17,6 +17,7 @@ set(gtest_bump_tests
     bump_coordset_extents.cpp
     bump_make_polyhedral_topology.cpp
     bump_mergemeshes.cpp
+    bump_mesh_operations.cpp
     bump_node_to_zone_relation.cpp
     bump_slicers.cpp
     bump_topology_mapper.cpp
